Close ZIP once finished
It is important to close the ZIP once data is finished being written to
the WACZ or else some of the data may not be flushed to disk. This is
probably more important for usage of py-wacz as a library since the file
would automatically get flushed when it is used from the command line.

Fixes #20